requirements.txt
langchain
faiss-cpu
# faiss-gpu
llama-cpp-python
numpy
requests
sentence_transformers
faker
streamlit